pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.37",
    "flask-migrate>=4.1.0",
    "oauthlib>=3.2.2",
    "openai>=1.61.1",
    "redis>=5.2.1",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.160.0",
]
